requirements.txt
langchain
langchain-core
langchain-openai
langchain-groq
openai
groq
python-dotenv
